Renamed javafxapplication1
Changed package and application name from javafxapplication1 to userInterface and Game
