+++ Cluedo/nbproject/project.properties
@@ -80,7 +80,7 @@ javafx.enable.concurrent.external.runs=false
 javafx.enabled=true
 javafx.fallback.class=com.javafx.main.NoJavaFXFallback
 # Main class for JavaFX
-javafx.main.class=javafxapplication1.JavaFXApplication1
+javafx.main.class=userInterface.Game
 javafx.preloader.class=
 # This project does not use Preloader
 javafx.preloader.enabled=false
